refactor(mobile): 역 검색 api repository 경계 분리
