+++ apps/mobile/lib/app/app_dependencies.dart
@@ -7,6 +7,7 @@ import '../favorite_facility.dart';
 import '../features/favorites/data/drift_favorite_repositories.dart';
 import '../features/preferences/data/drift_notification_settings_repository.dart';
 import '../features/search_history/data/drift_search_history_repository.dart';
+import '../features/stations/data/station_api_repository.dart';
 import '../features/stations/data/drift_station_repository.dart';
 import '../internal_route.dart';
 import '../notification_settings.dart';
